backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
